+++ src/conversions/neon/interpolator.rs
@@ -229,23 +229,17 @@ impl<const GRID_SIZE: usize> Fetcher<NeonVectorDouble>
 
 pub(crate) trait NeonMdInterpolation<'a, const GRID_SIZE: usize> {
     fn new(table: &'a [NeonAlignedF32]) -> Self;
-    fn inter3_neon(&self, in_r: u16, in_g: u16, in_b: u16) -> NeonVector;
+    fn inter3_neon(&self, in_r: u8, in_g: u8, in_b: u8) -> NeonVector;
 }
 
 pub(crate) trait NeonMdInterpolationDouble<'a, const GRID_SIZE: usize> {
     fn new(table0: &'a [NeonAlignedF32], table1: &'a [NeonAlignedF32]) -> Self;
-    fn inter3_neon(&self, in_r: u16, in_g: u16, in_b: u16) -> (NeonVector, NeonVector);
+    fn inter3_neon(&self, in_r: u8, in_g: u8, in_b: u8) -> (NeonVector, NeonVector);
 }
 
 impl<const GRID_SIZE: usize> TetrahedralNeon<'_, GRID_SIZE> {
     #[inline(always)]
-    fn interpolate(
-        &self,
-        in_r: u16,
-        in_g: u16,
-        in_b: u16,
-        r: impl Fetcher<NeonVector>,
-    ) -> NeonVector {
+    fn interpolate(&self, in_r: u8, in_g: u8, in_b: u8, r: impl Fetcher<NeonVector>) -> NeonVector {
         const SCALE: f32 = 1.0 / LUT_SAMPLING as f32;
         let x: i32 = in_r as i32 * (GRID_SIZE as i32 - 1) / LUT_SAMPLING as i32;
         let y: i32 = in_g as i32 * (GRID_SIZE as i32 - 1) / LUT_SAMPLING as i32;
@@ -309,9 +303,9 @@ impl<const GRID_SIZE: usize> TetrahedralNeonDouble<'_, GRID_SIZE> {
     #[inline(always)]
     fn interpolate(
         &self,
-        in_r: u16,
-        in_g: u16,
-        in_b: u16,
+        in_r: u8,
+        in_g: u8,
+        in_b: u8,
         r: impl Fetcher<NeonVectorDouble>,
     ) -> (NeonVector, NeonVector) {
         const SCALE: f32 = 1.0 / LUT_SAMPLING as f32;
@@ -384,7 +378,7 @@ macro_rules! define_md_inter_neon {
             }
 
             #[inline(always)]
-            fn inter3_neon(&self, in_r: u16, in_g: u16, in_b: u16) -> NeonVector {
+            fn inter3_neon(&self, in_r: u8, in_g: u8, in_b: u8) -> NeonVector {
                 self.interpolate(
                     in_r,
                     in_g,
@@ -410,7 +404,7 @@ macro_rules! define_md_inter_neon_d {
             }
 
             #[inline(always)]
-            fn inter3_neon(&self, in_r: u16, in_g: u16, in_b: u16) -> (NeonVector, NeonVector) {
+            fn inter3_neon(&self, in_r: u8, in_g: u8, in_b: u8) -> (NeonVector, NeonVector) {
                 self.interpolate(
                     in_r,
                     in_g,
@@ -436,13 +430,7 @@ define_md_inter_neon_d!(TrilinearNeonDouble);
 
 impl<const GRID_SIZE: usize> PyramidalNeon<'_, GRID_SIZE> {
     #[inline(always)]
-    fn interpolate(
-        &self,
-        in_r: u16,
-        in_g: u16,
-        in_b: u16,
-        r: impl Fetcher<NeonVector>,
-    ) -> NeonVector {
+    fn interpolate(&self, in_r: u8, in_g: u8, in_b: u8, r: impl Fetcher<NeonVector>) -> NeonVector {
         const SCALE: f32 = 1.0 / LUT_SAMPLING as f32;
         let x: i32 = in_r as i32 * (GRID_SIZE as i32 - 1) / LUT_SAMPLING as i32;
         let y: i32 = in_g as i32 * (GRID_SIZE as i32 - 1) / LUT_SAMPLING as i32;
@@ -513,9 +501,9 @@ impl<const GRID_SIZE: usize> PyramidalNeonDouble<'_, GRID_SIZE> {
     #[inline(always)]
     fn interpolate(
         &self,
-        in_r: u16,
-        in_g: u16,
-        in_b: u16,
+        in_r: u8,
+        in_g: u8,
+        in_b: u8,
         r: impl Fetcher<NeonVectorDouble>,
     ) -> (NeonVector, NeonVector) {
         const SCALE: f32 = 1.0 / LUT_SAMPLING as f32;
@@ -596,13 +584,7 @@ impl<const GRID_SIZE: usize> PyramidalNeonDouble<'_, GRID_SIZE> {
 
 impl<const GRID_SIZE: usize> PrismaticNeon<'_, GRID_SIZE> {
     #[inline(always)]
-    fn interpolate(
-        &self,
-        in_r: u16,
-        in_g: u16,
-        in_b: u16,
-        r: impl Fetcher<NeonVector>,
-    ) -> NeonVector {
+    fn interpolate(&self, in_r: u8, in_g: u8, in_b: u8, r: impl Fetcher<NeonVector>) -> NeonVector {
         const SCALE: f32 = 1.0 / LUT_SAMPLING as f32;
         let x: i32 = in_r as i32 * (GRID_SIZE as i32 - 1) / LUT_SAMPLING as i32;
         let y: i32 = in_g as i32 * (GRID_SIZE as i32 - 1) / LUT_SAMPLING as i32;
@@ -664,9 +646,9 @@ impl<const GRID_SIZE: usize> PrismaticNeonDouble<'_, GRID_SIZE> {
     #[inline(always)]
     fn interpolate(
         &self,
-        in_r: u16,
-        in_g: u16,
-        in_b: u16,
+        in_r: u8,
+        in_g: u8,
+        in_b: u8,
         rv: impl Fetcher<NeonVectorDouble>,
     ) -> (NeonVector, NeonVector) {
         const SCALE: f32 = 1.0 / LUT_SAMPLING as f32;
@@ -736,9 +718,9 @@ impl<const GRID_SIZE: usize> TrilinearNeonDouble<'_, GRID_SIZE> {
     #[inline(always)]
     fn interpolate(
         &self,
-        in_r: u16,
-        in_g: u16,
-        in_b: u16,
+        in_r: u8,
+        in_g: u8,
+        in_b: u8,
         r: impl Fetcher<NeonVectorDouble>,
     ) -> (NeonVector, NeonVector) {
         const SCALE: f32 = 1.0 / LUT_SAMPLING as f32;
@@ -790,13 +772,7 @@ impl<const GRID_SIZE: usize> TrilinearNeonDouble<'_, GRID_SIZE> {
 
 impl<const GRID_SIZE: usize> TrilinearNeon<'_, GRID_SIZE> {
     #[inline(always)]
-    fn interpolate(
-        &self,
-        in_r: u16,
-        in_g: u16,
-        in_b: u16,
-        r: impl Fetcher<NeonVector>,
-    ) -> NeonVector {
+    fn interpolate(&self, in_r: u8, in_g: u8, in_b: u8, r: impl Fetcher<NeonVector>) -> NeonVector {
         const SCALE: f32 = 1.0 / LUT_SAMPLING as f32;
         let x: i32 = in_r as i32 * (GRID_SIZE as i32 - 1) / LUT_SAMPLING as i32;
         let y: i32 = in_g as i32 * (GRID_SIZE as i32 - 1) / LUT_SAMPLING as i32;
